Documentação com Swagger UI.

+++ BancoEisen.AuthProvider/Controllers/UsuariosController.cs
@@ -1,5 +1,6 @@
 ﻿using BancoEisen.AuthProvider.Models;
 using BancoEisen.AuthProvider.Services;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
@@ -9,6 +10,11 @@ namespace BancoEisen.AuthProvider.Controllers
     [ApiController]
     [ApiVersion("1.0")]
     [Route("api/v{version:apiVersion}/[controller]")]
+    [ApiExplorerSettings(GroupName = "v1.0")]
+    [Consumes("application/json", "text/json")]
+    [Produces("application/json", "text/json", "application/xml", "text/xml")]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(ErrorResponse))]
     public class UsuariosController : ControllerBase
     {
         private readonly IUsuarioService usuarioService;
@@ -18,8 +24,13 @@ namespace BancoEisen.AuthProvider.Controllers
             this.usuarioService = usuarioService;
         }
 
+        /// <summary>
+        /// Solicita o envio de e-mail contendo o token para redefinição de e-mail.
+        /// </summary>
         [HttpGet("solicitarAlteracaoEmail")]
-        public async Task<IActionResult> SolicitarRedefinicaoSenha([Required][FromQuery] string nomeUsuario,
+        [ProducesResponseType(StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<IActionResult> SolicitarRedefinicaoEmail([Required][FromQuery] string nomeUsuario,
                                                                    [Required][FromQuery] string novoEmail)
         {
             if (!ModelState.IsValid)
@@ -35,7 +46,13 @@ namespace BancoEisen.AuthProvider.Controllers
             return Accepted();
         }
 
+        /// <summary>
+        /// Solicita o envio de e-mail contendo o token para redefinição de senha.
+        /// </summary>
         [HttpGet("solicitarRedefinicaoSenha")]
+        [ProducesResponseType(StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<IActionResult> SolicitarRedefinicaoSenha([Required][FromQuery] string nomeUsuario)
         {
             if (!ModelState.IsValid)
@@ -51,7 +68,11 @@ namespace BancoEisen.AuthProvider.Controllers
             return Accepted();
         }
 
+        /// <summary>
+        /// Realiza a criação do usuário.
+        /// </summary>
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
         public async Task<IActionResult> Cadastrar(UsuarioInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -62,7 +83,7 @@ namespace BancoEisen.AuthProvider.Controllers
             if (result != null)
             {
                 if (!result.Succeeded)
-                    return BadRequest(result.Errors);
+                    return BadRequest(ErrorResponse.From(result.Errors));
 
                 var usuario = await usuarioService.ObterPeloNomeAsync(informacoes.NomeUsuario);
 
@@ -74,7 +95,12 @@ namespace BancoEisen.AuthProvider.Controllers
             return Created(loginUri, null);
         }
 
+        /// <summary>
+        /// Realiza a confirmação de e-mail.
+        /// </summary>
         [HttpPatch]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> ConfirmarEmail(ConfirmacaoInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -93,7 +119,12 @@ namespace BancoEisen.AuthProvider.Controllers
             return Ok();
         }
 
+        /// <summary>
+        /// Realiza a alteração de e-mail.
+        /// </summary>
         [HttpPut("alterarEmail")]
+        [ProducesResponseType(StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> AlterarEmail(AlteracaoEmailInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -119,7 +150,12 @@ namespace BancoEisen.AuthProvider.Controllers
             return Accepted();
         }
 
+        /// <summary>
+        /// Realiza a alteração de senha.
+        /// </summary>
         [HttpPut("alterarSenha")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> AlterarSenha(AlteracaoSenhaInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -138,7 +174,12 @@ namespace BancoEisen.AuthProvider.Controllers
             return Ok();
         }
 
+        /// <summary>
+        /// Define se o usuário irá ou não utilizar autenticação de dois fatores.
+        /// </summary>
         [HttpPut("alterarAutenticacaoDoisFatores")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> AlterarAutenticacaoDoisFatores(AlterarAutenticacaoDoisFatoresInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -157,7 +198,12 @@ namespace BancoEisen.AuthProvider.Controllers
             return Ok();
         }
 
+        /// <summary>
+        /// Realiza a alteração de senha através do token.
+        /// </summary>
         [HttpPut("redefinirSenha")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> RedefinirSenha(RedefinicaoSenhaInformacoes informacoes)
         {
             if (!ModelState.IsValid)
@@ -176,10 +222,18 @@ namespace BancoEisen.AuthProvider.Controllers
             return Ok();
         }
 
+        /// <summary>
+        /// Realiza a exclusão do usuário com o nome informado.
+        /// </summary>
         [HttpDelete("{nomeUsuario}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> Remover(string nomeUsuario)
         {
             var usuario = await usuarioService.ObterPeloNomeAsync(nomeUsuario);
+
+            if (usuario == null)
+                return NotFound();
 
             var result = await usuarioService.RemoverAsync(usuario);
